example/api/requirements.txt
coreapi==2.3.0
Django==1.10.4
django-crispy-forms==1.6.1
djangorestframework==3.6.2
djangorestframework-filters==0.9.1
Markdown==2.6.8
